Rename file to "decorations"
It no longer contains just the trailing underscore.
